+++ admin/includes/view_all_users.php
@@ -43,11 +43,10 @@
   echo "<td><a href='users.php?switch_to_admin={$user_id}'>Admin</a></td>";
   echo "<td><a href='users.php?switch_to_employee={$user_id}'>Employee</a></td>";
   echo "<td><a href='users.php?source=edit_user&edit_user={$user_id}'>Edit</a></td>";
-  echo "<td><a href='users.php?delete={$user_id}'>Delete</a></td>";
+  echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?'); \" href='users.php?delete={$user_id}'>Delete</a></td>";
   echo "</tr>";
   }
 ?>
-
 
 </tr>
 </tbody>
